refactor: 폴더 구조에 맞춰 GetPayrollCorrections.js 파일 이동
