+++ src/features/admin/approval/components/Table.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { getAllPayrollCorrections } from '@/shared/api/firebase/services/payrollCorrectionsService';
+import { getAllPayrollCorrections } from '@/features/admin/approval/api/GetPayrollCorrections';
 import { TableHeader } from './TableHeader';
 import { TableRow } from './TableRow';
 import { ExpandedRow } from './ExpandedRow';
